% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-rle.R
\name{coerce-rle}
\alias{coerce-rle}
\alias{asRle}
\alias{asRleDataFrame}
\alias{asVectorDataFrame}
\title{Tools to reduce memory size of DataFrames}
\usage{
asRle(x)

asRleDataFrame(x, columns = character())

asVectorDataFrame(x)
}
\arguments{
\item{x}{For \code{asRle}: \code{numeric}, \code{logical} or \code{character} vector to convert.}

\item{columns}{For \code{asRleDataFrame}: \code{character} with optional column names
that should always be converted to type \code{Rle}.}
}
\description{
These functions help reducing the memory requirement of \code{\link[=DataFrame]{DataFrame()}} objects
by converting columns with constant values to type \code{\link[=Rle]{Rle()}}.
\itemize{
\item \code{asRle}: convert a \code{numeric}, \code{logical} or \code{character} vector into an
\code{\link[=Rle]{Rle()}} if it will use less memory than the bare vector. Input arguments of
a different type are returned as-is.
\item \code{asRleDataFrame}: convert columns in \code{x} (a \code{\link[=DataFrame]{DataFrame()}}) to type \code{Rle}
(with \code{asRle}). Argument \code{columns} allows to specify columns that \strong{have}
to be converted (even if it would not reduce the memory demand).
\item \code{asVectorDataFrame}: convert all columns of type \code{Rle} in \code{x} back to base
vectors (using \code{as.vector}).
}
}
\seealso{
Other coerce functions: \code{\link{coerce}}
}
\author{
Johannes Rainer, Sebastian Gibb
}
\concept{coerce functions}
