% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{coerce}
\alias{coerce}
\alias{asInteger}
\title{Coerce functions}
\usage{
asInteger(x)
}
\arguments{
\item{x}{input argument.}
}
\description{
\itemize{
\item \code{asInteger}: convert \code{x} to an \code{integer} and throw an error if \code{x} is not
a \code{numeric}.
}
}
\examples{
## Convert numeric to integer
asInteger(3.4)

asInteger(3)
}
\author{
Johannes Rainer
}
\concept{coerce functions}
