% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralMining.R
\name{containsNeutralLoss}
\alias{containsNeutralLoss}
\title{Check if a certain neutral loss occured}
\usage{
containsNeutralLoss(
  x,
  precursorMz = 0,
  neutralLoss = 0,
  tolerance = 0,
  ppm = 0
)
}
\arguments{
\item{x}{\code{matrix} with two column where one contains m/z values
(column \code{"mz"}) and the second corresponding intensity values
(column \code{"intensity"})}

\item{precursorMz}{\code{numeric(1)} Precursor m/z from which the neutral loss is
occuring}

\item{neutralLoss}{\code{numeric(1)} Mass of neutral loss. Will be subtracted from
\code{precursorMz}}

\item{tolerance}{\code{numeric(1)}, accepted tolerance.}

\item{ppm}{\code{numeric(1)} representing a relative, value-specific
parts-per-million (PPM) tolerance that is added to \code{tolerance}.}
}
\value{
\code{logical} Returns either TRUE or FALSE, if fitting fragment m/z was
found or not
}
\description{
This checks if m/z value in a fragmentation spectrum is similar to a m/z
obtained from a neutral loss. The product ion m/z is calculated from the
arguments \code{precursorMz} and \code{neutralLoss}
}
\examples{
x <- matrix(c(c(95.08628, 97.10201, 123.11601, 240.26794, 252.26850,
               270.27852, 590.58606, 608.59827, 609.60059, 626.60829,
               627.61158, 628.61115),
             c(344, 454, 374, 268, 1206,
               996, 256, 2032, 1026, 7270,
               2386, 298)),  ncol = 2,)
               
colnames(x) <- c("mz", "intensity")

containsNeutralLoss(x,
                   precursorMz = 626.60816, neutralLoss = 18.010565,
                   tolerance = 0.005)
}
\seealso{
Other Spectral analysis: 
\code{\link{containsProductIon}()}
}
\author{
Michael Witting
}
\concept{Spectral analysis}
