% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralMining.R
\name{containsProductIon}
\alias{containsProductIon}
\title{Check if a certain specified product ions can be found}
\usage{
containsProductIon(x, productMz = 0, tolerance = 0, ppm = 0)
}
\arguments{
\item{x}{\code{matrix} with two column where one contains m/z values
(column \code{"mz"}) and the second corresponding intensity values
(column \code{"intensity"})}

\item{productMz}{\code{numeric} One or multiple fragment m/z that shall be present}

\item{tolerance}{\code{numeric(1)}, accepted tolerance.}

\item{ppm}{\code{numeric(1)} representing a relative, value-specific
parts-per-million (PPM) tolerance that is added to \code{tolerance}.}
}
\value{
\code{logical} Returns either TRUE or FALSE, if fitting fragment m/z was
found or not
}
\description{
This checks if one or multiple fragment m/z values are present in a
fragmentation spectrum
}
\examples{
x <- matrix(c(c(238.25334, 250.25265, 251.25765, 268.26535, 606.58326,
                607.58785, 624.59306, 625.59769, 626.60180),
              c(1088, 3852, 212, 200, 2178,
                724, 1266, 584, 222)),  ncol = 2,)
                
colnames(x) <- c("mz", "intensity")

result <- containsProductIon(x,
                            productMz = c(250.2529, 268.2635, 238.2530),
                            tolerance = 0.005)
}
\seealso{
Other Spectral analysis: 
\code{\link{containsNeutralLoss}()}
}
\author{
Michael Witting
}
\concept{Spectral analysis}
