% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotproduct.R
\name{dotproduct}
\alias{dotproduct}
\title{Calculate the normalized dot product}
\usage{
dotproduct(x, y, m = 0.5, n = 0)
}
\arguments{
\item{x}{\code{list}/\code{data.frame} of length 2 with m/z (\code{"mz"}) and corresponding
intensity values (\code{"intensity"})}

\item{y}{\code{list}/\code{data.frame} of length 2 with m/z (\code{"mz"}) and corresponding
intensity values (\code{"intensity"})}

\item{m}{\code{numeric(1)}, exponent for peak intensity-based weights}

\item{n}{\code{numeric(1)}, exponent for m/z-based weights}
}
\value{
\code{numeric(1)}, \code{dotproduct} returns a numeric similarity coefficient between
0 and 1.
}
\description{
Calculate the normalized dot product (NDP).\code{dotproduct} returns a numeric
value ranging between 0 and 1, where 0 indicates no similarity between the
two MS/MS features, while 1 indicates that the MS/MS features are identical.
}
\details{
Each row in \code{x} corresponds to the respective row in \code{y}, i.e. the peaks
(entries \code{"mz"}) per spectrum have to match.
\code{m} and \code{n} are weights given on the peak intensity and the m/z values
respectively. As default (\code{m = 0.5}), the square roots of the intensity
values are taken to calculate weights. With increasing values for \code{m}, high
intensity values will be taken more into account for similarity calculation,
i.e. differences between intensities will be intensified.
With increasing values for \code{n}, high m/z values will be taken more into
account for similarity calculation. Especially when working with small
molecules, a value \code{n > 0} can be set, to give a weight on the m/z values to
accommodate that shared fragments with higher m/z are less likely and will
mean that molecules might be more similar. If \code{n != 0}, a warning will be
raised if the corresponding m/z values are not identical, since small
differences in m/z values will distort the similarity values with increasing
\code{n}. If \code{m=0} or \code{n=0}, intensity values or m/z values, respectively, are not
taken into account.

The normalized dot product is calculated according to:
\deqn{NDP = \frac{\sum(W_{S1, i} \cdot W_{S2, i}) ^ 2}{ \sum(W_{S1, i} ^ 2) * \sum(W_{S2, i} ^ 2) }}{\sum(W_{S1, i} \cdot W_{S2, i}) ^ 2 \sum(W_{S1, i} ^ 2) * \sum(W_{S2, i} ^ 2)},
with \eqn{W = [ peak intensity] ^{m} \cdot [m/z]^n}.
For further information on normalized dot product see for example
Li et al. (2015).
Prior to calculating \deqn{W_{S1}} or \deqn{W_{S2}}, all intensity values
are divided by the maximum intensity value.
}
\examples{
x <- data.frame(mz=c(100.002, 100.001, NA, 300.01, 300.02, NA), 
        intensity=c(2, 1.5, 0, 1.2, 0.9, 0))
y <- data.frame(mz = c(100.0, NA, 200.0, 300.002, 300.025, 300.0255),
        intensity = c(2, 0, 3, 1, 4, 0.4))
dotproduct(x, y, m = 0.5, n = 0) 

}
\references{
Li et al. (2015): Navigating natural variation in herbivory-induced
secondary metabolism in coyote tobacco populations using MS/MS structural
analysis. PNAS, E4147--E4155.
}
\author{
Thomas Naake, \email{thomasnaake@googlemail.com}
}
