% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatRt.R
\name{formatRt}
\alias{formatRt}
\title{Format Retention Time}
\usage{
formatRt(rt)
}
\arguments{
\item{rt}{The retention time, either a \code{numeric()} in seconds or a
\code{character()} as \code{"mm:ss"}.}
}
\value{
A reformatted retention time.
}
\description{
Converts seconds to/from "min:sec" format.
}
\examples{
formatRt(1524)
formatRt(1)
formatRt("25:24")
}
\author{
Laurent Gatto
}
