% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPeaks.R
\name{graphPeaks}
\alias{graphPeaks}
\title{Match two spectra using bipartite networks and combinatorics}
\usage{
graphPeaks(x, y, ppm = 20, FUN = dotproduct, ...)
}
\arguments{
\item{x}{\code{matrix}, the first column (\code{"mz"}) contains m/z value and the
second column (\code{"intensity"}) contains the corresponding intensity values}

\item{y}{\code{matrix}, the first column (\code{"mz"}) contains m/z value and the
second column (\code{"intensity"}) contains the corresponding intensity values}

\item{ppm}{\code{numeric}, tolerance parameter in ppm to match corresponding peaks}

\item{FUN}{function to calculate similarity between spectra}

\item{...}{additional parameters passed to \code{FUN}}
}
\value{
list with elements \code{x} and \code{y} each being a matrix with columns \code{"mz"}
and \code{"intensity"}. Each row (peak) in \code{x} matches the row (peak) in \code{y}
}
\description{
Function to match spectrum objects using bipartite networks and combinations
\code{y} as input that contain spectral information. The matching
is a multi-step procedure:
\enumerate{
\item filtering based on \code{ppm},
\item retain order of matches between features of \code{x} and
\code{y} (remove crossing edges that violate the order of matching
m/z),
\item calculate all combinations of the remaining possibilities.
}
}
\details{
Objective function is highest similarites between the two
spectral objects, i.e. \code{FUN} is calculated over all combinations and
the similarity of the combination that yields the highest similarity is
returned.
}
\examples{
x <- matrix(c(c(100.001, 100.002, 300.01, 300.02),
    c(1, 1, 1, 1)), ncol = 2, nrow = 4, byrow = FALSE)
colnames(x) <- c("mz", "intensity")

y <- matrix(c(c(100.0, 200.0, 300.002, 300.025, 300.0255),
    c(1, 1, 1, 1, 1)), ncol = 2, nrow = 5, byrow = FALSE)
colnames(y) <- c("mz", "intensity")
graphPeaks(x = x, y = y, ppm = 20, FUN = dotproduct)
}
\author{
Thomas Naake \email{thomasnaake@googlemail.com}
}
