% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group_mz_int}
\alias{group_mz_int}
\title{Grouping of numeric values by similarity of m/z and intensity}
\usage{
group_mz_int(x, y = numeric(), max_num = Inf, tolerance = 0, ppm = 0)
}
\arguments{
\item{x}{\code{numeric} with the \emph{m/z} values to be grouped.}

\item{y}{\code{numeric} with the intensity values of the peaks.}

\item{max_num}{\code{integer(1)} defining the maximum number of peaks for a peak
group.}

\item{tolerance}{\code{numeric(1)} with the maximal accepted difference between
values in \code{x} to be grouped into the same entity.}

\item{ppm}{\code{numeric(1)} defining a value-dependent maximal accepted
difference between values in \code{x} expressed in parts-per-million.}
}
\value{
\code{integer} of length equal to \code{x} with the groups.
}
\description{
The \code{group_mz_int()} function groups peaks with similar m/z (across several
scans) considering also their intensity. The algorithm first orders peak
decreasingly by their intensity. Then it iteratively selects the peak with
the highest intensity that is not yet part of a peak group and finds all
other peaks with a difference in their m/z that is smaller than defined by
\code{tolerance} and \code{ppm}. These peaks are assigned to the same peak group.

Setting parameter \code{max_num} to a finite number forces each peak group to
contain only the at most \code{max_num} peaks ordered by their intensity.
}
\note{
This method solves the scenario like the difference between the smallest and
largest value in a group can be larger than \code{tolerance} and \code{ppm}.
}
\examples{

## Define a (sorted) numeric vector
x = c(56, 56.004, 56.008, 56.012, 56.016, 56.02)
y = c(52151, 125584, 582, 58452, 458, 57452)
max_num = 2

## With `ppm = 0` and `tolerance = 0` only identical values are grouped
group_mz_int(x, y, max_num)

## With `tolerance = 0.005`
group_mz_int(x, y, max_num, tolerance = 0.005)

## three groups were made.

## With ppm
group_mz_int(x, y, max_num, ppm = 10)

## Same on an unsorted vector
x <- c(56, 56.012, 56.016, 56.004, 56.008, 56.02)
y = c(52151, 58452, 458, 125584, 582, 57452)
group_mz_int(x, y, max_num, tolerance = 0.005)

## the same three groups were made.

}
\seealso{
\code{\link[=group]{group()}}
}
\author{
Muyao Xi
}
