% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logMessage.R
\name{logMessage}
\alias{logMessage}
\title{Create Log Messages}
\usage{
logMessage(x, ..., date = TRUE)
}
\arguments{
\item{x}{\code{character}, existing log messages.}

\item{...}{one or more object to be converted to \code{character}, the new log
message.}

\item{date}{\code{logical(1)}, should the \code{\link[=date]{date()}} append to the message?}
}
\value{
\code{character} vector with new log message appended.
}
\description{
This function creates a standardised log message and append it to an
existing vector of log messages.
}
\examples{
logMessage(NULL, "Important Stuff")
logMessage(NULL, "Important Stuff", date = FALSE)
logMessage("Old message",
           "New number: ", 3, " and letters ",
           paste0(letters[1:4], collapse=", "))
}
\author{
Sebastian Gibb
}
