% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeMethods}
\alias{normalizeMethods}
\alias{normalize_matrix}
\title{Quantitative data normalisation}
\usage{
normalizeMethods()

normalize_matrix(x, method, ...)
}
\arguments{
\item{x}{A matrix to be normalised.}

\item{method}{\code{character(1)} defining the normalisation
method. See \code{normalizeMethods()} for available ones.}

\item{...}{Additional parameters passed to the inner normalisation
function.}
}
\value{
A normalised matrix of dimensions \code{dim(x)}.
}
\description{
Function to normalise a matrix of quantitative omics data. The
nature of the normalisation is controlled by the \code{method}
argument, described below.
}
\details{
The \code{method} parameter in \code{normalize} can be one of \code{"sum"},
\code{"max"}, \code{"quantiles"}, \code{"center.mean"}, \code{"center.median"},
\code{"center.median"}, \verb{"quantiles.robust}" or \code{"vsn"}.  For \code{"sum"}
and \code{"max"}, each feature's intensity is divided by the maximum or
the sum of the feature respectively. These two methods are applied
along the features (rows). The \code{normalizeMethods()} function
returns a vector of available normalisation methods.

\code{"center.mean"} and \code{"center.median"} translate the respective
sample (column) intensities according to the column mean or
median. \code{"diff.median"} translates all samples (columns) so that
they all match the grand median. Using \code{"quantiles"} or
\code{"quantiles.robust"} applies (robust) quantile normalisation, as
implemented in \code{\link[preprocessCore:normalize.quantiles]{preprocessCore::normalize.quantiles()}} and
\code{\link[preprocessCore:normalize.quantiles.robust]{preprocessCore::normalize.quantiles.robust()}}. \code{"vsn"} uses the
\code{\link[vsn:vsn2]{vsn::vsn2()}} function.  Note that the latter also glog-transforms
the intensities.  See respective manuals for more details and
function arguments.
}
\examples{
## test data
set.seed(42)
m <- matrix(rlnorm(60), 10)

normalize_matrix(m, method = "sum")

normalize_matrix(m, method = "max")

normalize_matrix(m, method = "quantiles")

normalize_matrix(m, method = "center.mean")
}
\author{
Laurent Gatto
}
