% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm}
\alias{ppm}
\title{PPM - Parts per Million}
\usage{
ppm(x, ppm)
}
\arguments{
\item{x}{`numeric`, mz values.}

\item{ppm}{`numeric`, parts per million (ppm) values.}
}
\value{
`numeric`: `x * 1e-6`
}
\description{
`ppm` is a small helper function to mulitply *mz* values with a given *ppm*.
}
\examples{
ppm(c(1000, 2000), 5)
}
\author{
Sebastian Gibb
}
