% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{between}
\alias{between}
\alias{\%between\%}
\title{Range helper functions}
\usage{
between(x, range)

x \%between\% range
}
\arguments{
\item{x}{\code{numeric}, input values.}

\item{range}{\code{numeric(2)}, range to compare against.}
}
\value{
\code{logical} vector of length \code{length(x)}.
}
\description{
These functions help to work with \code{numeric} ranges.
}
\examples{
between(1:4, 2:3)
1:4 \%between\% 2:3
}
\seealso{
Other helper functions for developers:
\code{\link[=isPeaksMatrix]{isPeaksMatrix()}},
\code{\link[=rbindFill]{rbindFill()}},
\code{\link[=validPeaksMatrix]{validPeaksMatrix()}},
\code{\link[=vapply1c]{vapply1c()}},
\code{\link[=which.first]{which.first()}}
}
\author{
Sebastian Gibb
}
\concept{helper functions for developers}
