% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphPeaks.R
\name{shiftMatrix}
\alias{shiftMatrix}
\title{Shift columns of a matrix by n and set added columns to def}
\usage{
shiftMatrix(mat, x, n, def = NA)
}
\arguments{
\item{mat}{\code{matrix}}

\item{x}{\code{numeric}, col indices to shift}

\item{n}{\code{numeric(1)}, gives the number by how many positions the columns
\code{x} of \code{mat} are shifted}

\item{def}{\code{character(1)}/\code{numeric(1)}, replacement value for added columns}
}
\value{
matrix with all combinations of shifted rows, only returns the
columns \code{x} of \code{mat}
}
\description{
\code{shiftMatrix} shifts columns of a matrix by \code{n} and
sets the added columns to \code{def}.
}
\details{
helper function for \code{graphPeaks}
}
\examples{
mat <- matrix(letters[1:18], ncol = 6, nrow = 3)
x <- c(2, 4, 6)
shiftMatrix(mat = mat, x = x, n = 1, def = NA)
}
\author{
Thomas Naake \email{thomasnaake@googlemail.com}
}
