% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{asInteger}
\alias{rbindFill}
\title{General utility functions}
\usage{
asInteger(x)

rbindFill(...)
}
\arguments{
\item{x}{input argument.}

\item{...}{For \code{rbindFill}: 2 or more: \code{matrix}, \code{data.frame} or \code{DataFrame}.}
}
\description{
\itemize{
\item \code{asInteger}: convert \code{x} to an \code{integer} and throw an error if \code{x} is not
a \code{numeric}.
\item \code{rbindFill}: combines instances of \code{matrix}, \code{data.frame} or \code{DataFrame}
objects into a single instance adding eventually missing columns (filling
them with \code{NA}s).
}
}
\note{
\code{rbindFill} might not work if one of the columns contains S4classes.
}
\examples{

## Convert numeric to integer
asInteger(3.4)

asInteger(3)

## Combine matrices
a <- matrix(1:9, nrow = 3, ncol = 3)
colnames(a) <- c("a", "b", "c")
b <- matrix(1:12, nrow = 3, ncol = 4)
colnames(b) <- c("b", "a", "d", "e")
rbindFill(a, b)
rbindFill(b, a, b)
}
\author{
Johannes Rainer, Sebastian Gibb
}
