% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which.first}
\alias{which.first}
\alias{which.last}
\title{Which is the first/last TRUE value.}
\usage{
which.first(x)

which.last(x)
}
\arguments{
\item{x}{\code{logical}, vector.}
}
\value{
\code{integer}, index of the first/last \code{TRUE}
value. \code{integer(0)} if no \code{TRUE} (everything \code{FALSE} or \code{NA})
was found.
}
\description{
Determines the location, i.e., index of the first or last \code{TRUE}
value in a logical vector.
}
\examples{
l <- 2 <= 1:3
which.first(l)
which.last(l)
}
\seealso{
\code{\link[=which.min]{which.min()}}

Other helper functions for developers:
\code{\link[=between]{between()}},
\code{\link[=isPeaksMatrix]{isPeaksMatrix()}},
\code{\link[=rbindFill]{rbindFill()}},
\code{\link[=validPeaksMatrix]{validPeaksMatrix()}},
\code{\link[=vapply1c]{vapply1c()}}
}
\author{
Sebastian Gibb
}
\concept{helper functions for developers}
