% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin}
\alias{bin}
\title{Binning}
\usage{
bin(x, toBin, binSize = 1, breaks = seq(floor(min(toBin)),
  ceiling(max(toBin)), by = binSize), FUN = max)
}
\arguments{
\item{x}{`numeric` with the values that should be aggregated/binned.}

\item{toBin}{`numeric` with same length than `x` with values to be used for
the binning.}

\item{binSize}{`numeric(1)` with the size of a bin.}

\item{breaks}{`numeric` defining the breaks (bins).}

\item{FUN}{`function` to be used to aggregate values of `x` falling into the
bins defined by `breaks`.}
}
\value{
`list` with elements `x` (aggregated values of `x`) and `mids` (the
    bin mid points).
}
\description{
Aggregate values in `x` for bins defined on `toBin`: all values
in `x` for values in `toBin` falling into a bin (defined on `toBin`) are
aggregated with the provided function `FUN`.
}
\author{
Johannes Rainer, Sebastian Gibbals
}
