% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin}
\alias{bin}
\title{Binning}
\usage{
bin(x, y, size = 1, breaks = seq(floor(min(y)), ceiling(max(y)), by =
  size), FUN = max)
}
\arguments{
\item{x}{`numeric` with the values that should be aggregated/binned.}

\item{y}{`numeric` with same length than `x` with values to be used for
the binning.}

\item{size}{`numeric(1)` with the size of a bin.}

\item{breaks}{`numeric` defining the breaks (bins).}

\item{FUN}{`function` to be used to aggregate values of `x` falling into the
bins defined by `breaks`.}
}
\value{
`list` with elements `x` (aggregated values of `x`) and `mids` (the
    bin mid points).
}
\description{
Aggregate values in `x` for bins defined on `y`: all values
in `x` for values in `y` falling into a bin (defined on `y`) are
aggregated with the provided function `FUN`.
}
\examples{

## Define example intensities and m/z values
ints <- abs(rnorm(20, mean = 40))
mz <- seq(1:length(ints)) + rnorm(length(ints), sd = 0.001)

## Bin intensities by m/z bins with a bin size of 2
bin(ints, mz, size = 2)

## Repeat but summing up intensities instead of taking the max
bin(ints, mz, size = 2, FUN = sum)
}
\author{
Johannes Rainer, Sebastian Gibb
}
