% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{bin}
\alias{bin}
\title{Binning}
\usage{
bin(x, y, size = 1, breaks = seq(floor(min(y)), ceiling(max(y)), by =
  size), FUN = max)
}
\arguments{
\item{x}{\code{numeric} with the values that should be aggregated/binned.}

\item{y}{\code{numeric} with same length than \code{x} with values to be used for
the binning.}

\item{size}{\code{numeric(1)} with the size of a bin.}

\item{breaks}{\code{numeric} defining the breaks (bins).}

\item{FUN}{\code{function} to be used to aggregate values of \code{x} falling into the
bins defined by \code{breaks}.}
}
\value{
\code{list} with elements \code{x} (aggregated values of \code{x}) and \code{mids} (the
bin mid points).
}
\description{
Aggregate values in \code{x} for bins defined on \code{y}: all values
in \code{x} for values in \code{y} falling into a bin (defined on \code{y}) are
aggregated with the provided function \code{FUN}.
}
\examples{

## Define example intensities and m/z values
ints <- abs(rnorm(20, mean = 40))
mz <- seq(1:length(ints)) + rnorm(length(ints), sd = 0.001)

## Bin intensities by m/z bins with a bin size of 2
bin(ints, mz, size = 2)

## Repeat but summing up intensities instead of taking the max
bin(ints, mz, size = 2, FUN = sum)
}
\seealso{
Other grouping/matching functions: \code{\link{joinGraph}}
}
\author{
Johannes Rainer, Sebastian Gibb
}
\concept{grouping/matching functions}
