% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Spectral entropy}
\usage{
entropy(x)
}
\arguments{
\item{x}{\code{numeric}, intensities of the fragment ions.}
}
\value{
\code{numeric}: (normalized) entropy \code{x}.
}
\description{
These functions allow to calculate entropy measurements of an MS/MS
spectrum based on the metrics suggested by
Li et al. (https://doi.org/10.1038/s41592-021-01331-z).
Spectral entropy and normalized entropy are used to measure the complexity
of an spectra.
MassBank of North America (MoNA) defines spectra entropy as the intensity
weighted spectral peak number
(https://mona.fiehnlab.ucdavis.edu/documentation/entropy.
Additionally it is suggested to label as low-quality spectrums the ones
that have a normalized entropy higher than 0.8 and spectral entropy
higher than 3.
}
\examples{
spectrum <- rbind(c(41.04, 37.16), c(69.07, 66.83), c(86.1, 999.0))

entropy(spectrum[,2])
nentropy(spectrum[,2])

}
\author{
Mar Garcia-Aloy
}
\concept{helper functions for users

#' @references
Li, Y., Kind, T., Folz, J., Vaniya, A., Mehta. S.S., Fiehn, O. (2021).
Spectral entropy outperforms MS/MS dot product similarity for
small-molecule compound identification.
Nature Methods. 2021;18(12):1524-1531.
\doi{10.1038/s41592-021-01331-z}.}
