% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_sorted.R
\name{force_sorted}
\alias{force_sorted}
\title{Forcing a numeric vector into a monotonously increasing sequence.}
\usage{
force_sorted(x)
}
\arguments{
\item{x}{\code{numeric} vector.}
}
\value{
A vector with continuously increasing values.
}
\description{
This function performs interpolation on the non-increasing parts of an
input vector to ensure its values are continuously increasing.
}
\note{
NA values will not be replaced.
}
\examples{
x <- c(NA, NA, NA, 1.2, 1.1, 1.14, 1.2, 1.3, NA, 1.04, 1.4, 1.6, NA, NA)
sorted_rtime <- force_sorted(x)
is.unsorted(x, na.rm = TRUE)

}
