% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm}
\alias{ppm}
\title{PPM - Parts per Million}
\usage{
ppm(x)
}
\arguments{
\item{x}{`numeric`, ppm values.}
}
\value{
`numeric`: `x * 1e-6`
}
\description{
`ppm` is a small helper function to mulitply *ppm* values with `1e-6`.
}
\examples{
ppm(5)
}
\author{
Sebastian Gibb
}
