% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm}
\alias{ppm}
\title{PPM - Parts per Million}
\usage{
ppm(x, ppm)
}
\arguments{
\item{x}{`numeric`, value(s) used for ppm calculation, e.g. mz value(s).}

\item{ppm}{`numeric`, parts-per-million (ppm) value(s).}
}
\value{
`numeric`: parts-per-million of `x`
}
\description{
`ppm` is a small helper function to determine the parts-per-million for a
user-provided value and ppm.
}
\examples{
ppm(c(1000, 2000), 5)
}
\author{
Sebastian Gibb
}
