% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry expression on failure}
\usage{
retry(expr, ntimes = 5L, sleep_mult = 0L, retry_on = "*", ...)
}
\arguments{
\item{expr}{Expression to be evaluated.}

\item{ntimes}{\code{integer(1)} with the number of times to try.}

\item{sleep_mult}{\code{numeric(1)} multiplier to define the increasing waiting
time (in seconds).}

\item{retry_on}{\code{character(1)} pattern for the error message to retry \code{expr}.
Defaults to \code{retry_on = "*"} hence retrying \code{expr} on any error that
occurs. This allows to restrict retrying \code{expr} for specific cases,
such as temporary internet connection problems. The pattern defined by
\code{retry_on} is directly passed to the \code{grepl()} function.}

\item{...}{optional parameters passed to \code{grepl()}.}
}
\description{
\code{retry()} retries, upon failure, the evaluation of an expression \code{exp} for
\code{ntimes} times waiting an increasing amount of time
between tries, i.e., waiting for \code{Sys.sleep(i * sleep_mult)} seconds between
each try \code{i}. If \code{expr} fails for \code{ntimes} times an error will be thrown.
}
\note{
Warnings are suppressed.
}
\examples{

## In the example below content is read from a web page that might
## temporarily be offline. It is assumed that the error message contains
## the pattern `"temporarily"` in which case the call is repeated 7 times
## with an increasing interval between tries to allow the resource to
## become available again.
\dontrun{
    res <- retry(readLines("https://some-unreliable-web-content"),
                 retry_on = "temporarily", ntimes = 7L, sleep_mult = 10)
}
}
\author{
Johannes Rainer
}
