% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Smoothing}
\usage{
smooth(x, hws, method = c("MovingAverage", "WeightedMovingAverage",
  "SavitzkyGolay"), ...)
}
\arguments{
\item{x}{\code{numeric}, i.e. m/z values.}

\item{hws}{\code{integer(1)}, half window size, the resulting window reaches from
\code{(i - hws):(i + hws)}.}

\item{method}{\code{character(1)}, used method. Currently MovingAverage,
WeightedMovingAverage and SavitzkyGolay are supported.}

\item{...}{further arguments passed to \code{method}. Currently just \code{k} for
the \emph{polynomial order} in \code{"SavitzkyGolay"} is supported.}
}
\value{
A \code{numeric} of the same length as \code{x}.
}
\description{
This function smoothes a numeric vector.
}
\details{
\code{method = "WeightedMovingAverage"}: calculates a weighted moving average
with weights depending on the distance from the center
calculated as \code{1/2^abs(-hws:hws)} with the sum of all weigths normalized
to 1.

\code{method = "SavitzkyGolay"}: uses the Savitzky-Golay-Filter for smoothing. It
supports an additional argument \code{k} (\code{integer(1)}) that controls the order of
the polynomial (default: \code{k = 3}) used to calculated the coefficients.
}
\examples{
x <- c(1:10, 9:1)
plot(x, type = "b", pch = 20)
m  <- c("MovingAverage", "WeightedMovingAverage", "SavitzkyGolay")
for (i in seq_along(m)) {
    lines(smooth(x, hws = 3L, method = m[i]), col = i + 1L, pch = 20,
          type = "b")
}
legend("bottom", legend = c("x", m), pch = 20,
       col = c(1L, seq_along(m) + 1L))
}
\references{
A. Savitzky and M. J. Golay. 1964.
Smoothing and differentiation of data by simplified least squares procedures.
Analytical chemistry, 36(8), 1627-1639.

M. U. Bromba and H. Ziegler. 1981.
Application hints for Savitzky-Golay digital smoothing filters.
Analytical Chemistry, 53(11), 1583-1586.
}
\seealso{
Other noise estimation and smoothing functions: \code{\link{noise}}
}
\author{
Sebastian Gibb, Sigurdur Smarason (weighted moving average)
}
\concept{noise estimation and smoothing functions}
