% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys.R
\name{valleys}
\alias{valleys}
\title{Find Peak Valleys}
\usage{
valleys(x, p)
}
\arguments{
\item{x}{\code{numeric}, e.g. intensity values.}

\item{p}{\code{numeric}, indices of identified peaks/local maxima.}
}
\value{
A \code{matrix} with three columns representing the index of the left
valley, the peak center, and the right valley.
}
\description{
This function finds the valleys around peaks.
}
\examples{
ints <- c(5, 8, 12, 7, 4, 9, 15, 16, 11, 8, 3, 2, 3, 2, 9, 12, 14, 13, 8, 3)
mzs <- seq_along(ints)
peaks <- which(localMaxima(ints, hws = 3))
cols <- seq_len(peaks) + 1

plot(mzs, ints, type = "h", ylim = c(0, 16))
points(mzs[peaks], ints[peaks], col = cols, pch = 20)

v <- valleys(ints, peaks)
segments(mzs[v[, "left"]], 0, mzs[v[, "right"]], col = cols, lwd = 2)
}
\author{
Sebastian Gibb
}
