% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapply.R
\name{vapply1c}
\alias{vapply1c}
\alias{vapply1d}
\alias{vapply1l}
\title{vapply wrappers}
\usage{
vapply1c(X, FUN, ..., USE.NAMES = FALSE)

vapply1d(X, FUN, ..., USE.NAMES = FALSE)

vapply1l(X, FUN, ..., USE.NAMES = FALSE)
}
\arguments{
\item{X}{a vector (atomic or `list`).}

\item{FUN}{the `function` to be applied to each element of `X`.}

\item{...}{optional arguments to `FUN`.}

\item{USE.NAMES}{`logical`, should the return value be named.}
}
\value{
`vapply1c` returns a vector of `character`s of length `X`.

`vapply1d` returns a vector of `double`s of length `X`.

`vapply1l` returns a vector of `logical`s of length `X`.
}
\description{
These functions are short wrappers around typical `vapply` calls for easier
development.
}
\examples{
l <- list(a=1:3, b=4:6)
vapply1d(l, sum)
}
\author{
Sebastian Gibb
}
